+++ Realm/RLMUser.h
@@ -31,7 +31,7 @@
 typedef NS_ENUM(NSUInteger, RLMUserState) {
     /// The user is logged out. Call `logInWithCredentials:...` with valid credentials to log the user back in.
     RLMUserStateLoggedOut,
-    /// The user is logged in, and any Realms associated with it are syncing with Atlas Device Sync.
+    /// The user is logged in, and any Realms associated with it are syncing with Atlas App Services.
     RLMUserStateLoggedIn,
     /// The user has been removed, and cannot be used.
     RLMUserStateRemoved
@@ -56,7 +56,7 @@ NS_ASSUME_NONNULL_BEGIN
 
  A user may have one or more credentials associated with it. These credentials
  uniquely identify the user to the authentication provider, and are used to sign
- into a Atlas Device Sync user account.
+ into a Atlas App Services user account.
 
  Note that user objects are only vended out via SDK APIs, and cannot be directly
  initialized. User objects can be accessed from any thread.
@@ -64,7 +64,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface RLMUser : NSObject
 
 /**
- The unique Atlas Device Sync string identifying this user.
+ The unique Atlas App Services string identifying this user.
  Note this is different from an identitiy: A user may have multiple identities but has a single indentifier. See RLMUserIdentity.
  */
 @property (nonatomic, readonly) NSString *identifier NS_SWIFT_NAME(id);
@@ -83,7 +83,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  The user's refresh token used to access the Realm Application.
 
- This is required to make HTTP requests to Atlas Device Sync's REST API
+ This is required to make HTTP requests to Atlas App Services' REST API
  for functionality not exposed natively. It should be treated as sensitive data.
  */
 @property (nullable, nonatomic, readonly) NSString *accessToken;
@@ -229,9 +229,9 @@ NS_ASSUME_NONNULL_BEGIN
 - (RLMMongoClient *)mongoClientWithServiceName:(NSString *)serviceName NS_REFINED_FOR_SWIFT;
 
 /**
- Calls the Atlas Device Sync function with the provided name and arguments.
+ Calls the Atlas App Services function with the provided name and arguments.
 
- @param name The name of the Atlas Device Sync function to be called.
+ @param name The name of the Atlas App Services function to be called.
  @param arguments The `BSONArray` of arguments to be provided to the function.
  @param completion The completion handler to call when the function call is complete.
  This handler is executed on a non-main global `DispatchQueue`.